requirements.txt
Django==4.0.1
python-dotenv==0.19.2
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.8
